requirements.txt
asgiref==3.8.1
certifi==2025.4.26
chardet==4.0.0
Django==5.1.7
django-cors-headers==4.7.0
django-stubs==5.2.1
django-stubs-ext==5.2.1
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.2
djangorestframework-stubs==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==2.10
packaging==25.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
requests==2.25.1
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
types-PyYAML==6.0.12.20250516
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing_extensions==4.12.2
urllib3==1.26.20
django_informixdb
celery
redis==4.5.5
pyodbc
pytz
emoji